feat: resource aggregate feat: addons redesign optimize: change file name and object name
refactor: resolve code validation errors
test: add proxy aggregate test file
chore: add golingci-lint tool

+++ pkg/controller/resource-binding/common.go
@@ -31,6 +31,9 @@ func deleteClusterResource(ctx context.Context, clientSet client.Client, binding
 		return err
 	}
 	for _, item := range clusterResourceList.Items {
+		if !item.GetDeletionTimestamp().IsZero() {
+			continue
+		}
 		err = clientSet.Delete(ctx, &item)
 		if err != nil {
 			return err
